+++ src/psd_tools/api/composer.py
@@ -38,7 +38,7 @@ def intersect(*bboxes):
     return result
 
 
-def _blend(target, image, offset, mask=None):
+def _blend(target, image, offset):
     if offset[0] < 0:
         if image.width <= -offset[0]:
             return target
@@ -121,10 +121,11 @@ def compose(layers, bbox=None, layer_filter=None, color=None, **kwargs):
     # Alpha must be forced to correctly blend.
     mode = get_pil_mode(valid_layers[0]._psd.color_mode, True)
     result = Image.new(
-        mode, (bbox[2] - bbox[0], bbox[3] - bbox[1]), color=color,
+        mode, (bbox[2] - bbox[0], bbox[3] - bbox[1]),
+        color=color if color is not None else 'white',
     )
+    result.putalpha(0)
 
-    initial_layer = True
     for layer in valid_layers:
         if intersect(layer.bbox, bbox) == (0, 0, 0, 0):
             continue
@@ -135,11 +136,7 @@ def compose(layers, bbox=None, layer_filter=None, color=None, **kwargs):
 
         logger.debug('Composing %s' % layer)
         offset = (layer.left - bbox[0], layer.top - bbox[1])
-        if initial_layer:
-            result.paste(image, offset)
-            initial_layer = False
-        else:
-            result = _blend(result, image, offset)
+        result = _blend(result, image, offset)
 
     return result
 
@@ -151,7 +148,9 @@ def compose_layer(layer, force=False, **kwargs):
 
     image = layer.topil(**kwargs)
     if image is None or force:
-        image = create_fill(layer)
+        texture = create_fill(layer)
+        if texture is not None:
+            image = texture
 
     if image is None:
         return image
@@ -173,10 +172,12 @@ def compose_layer(layer, force=False, **kwargs):
                 # What should we do here? There are two alpha channels.
                 pass
             image.putalpha(mask)
-    elif layer.has_vector_mask() and not layer.has_pixels():
+    elif layer.has_vector_mask() and (force or not layer.has_pixels()):
         mask = draw_vector_mask(layer)
         # TODO: Stroke drawing.
-        image.putalpha(mask)
+        texture = image
+        image = Image.new(image.mode, image.size, 'white')
+        image.paste(texture, mask=mask)
 
     # Apply layer fill effects.
     apply_effect(layer, image)
@@ -210,16 +211,23 @@ def compose_layer(layer, force=False, **kwargs):
 def create_fill(layer):
     from PIL import Image
     mode = get_pil_mode(layer._psd.color_mode, True)
-    image = Image.new(mode, (layer.width, layer.height))
+    image = None
     if 'SOLID_COLOR_SHEET_SETTING' in layer.tagged_blocks:
+        image = Image.new(mode, (layer.width, layer.height), 'white')
         setting = layer.tagged_blocks.get_data(
             'SOLID_COLOR_SHEET_SETTING'
         )
-        draw_solid_color_fill(image, setting)
+        draw_solid_color_fill(image, setting, blend=False)
+    elif 'VECTOR_STROKE_CONTENT_DATA' in layer.tagged_blocks:
+        image = Image.new(mode, (layer.width, layer.height), 'white')
+        setting = layer.tagged_blocks.get_data('VECTOR_STROKE_CONTENT_DATA')
+        draw_solid_color_fill(image, setting, blend=False)
     elif 'PATTERN_FILL_SETTING' in layer.tagged_blocks:
+        image = Image.new(mode, (layer.width, layer.height), 'white')
         setting = layer.tagged_blocks.get_data('PATTERN_FILL_SETTING')
-        draw_pattern_fill(image, layer._psd, setting)
+        draw_pattern_fill(image, layer._psd, setting, blend=False)
     elif 'GRADIENT_FILL_SETTING' in layer.tagged_blocks:
+        image = Image.new(mode, (layer.width, layer.height), 'white')
         setting = layer.tagged_blocks.get_data('GRADIENT_FILL_SETTING')
         draw_gradient_fill(image, setting, blend=False)
     return image
@@ -257,42 +265,129 @@ def apply_effect(layer, image):
 
 
 def draw_vector_mask(layer):
-    from PIL import Image, ImageDraw
+    from PIL import Image, ImageChops
     width = layer._psd.width
     height = layer._psd.height
-    color = layer.vector_mask.initial_fill_rule * 255
-    mask = Image.new('L', (width, height), color)
-    draw = ImageDraw.Draw(mask)
-    for subpath in layer.vector_mask.paths:
-        path = [(
-            int(knot.anchor[1] * width),
-            int(knot.anchor[0] * height),
-        ) for knot in subpath]
-        # TODO: Use bezier curve instead of polygon. Perhaps aggdraw module.
-        draw.polygon(path, fill=(255 - color))
-    del draw
+    color = 255 * layer.vector_mask.initial_fill_rule
 
+    mask = Image.new('L', (width, height), color)
+    first = True
+    for subpath in layer.vector_mask.paths:
+        plane = _draw_subpath(subpath, width, height)
+        if subpath.operation == 0:
+            mask = ImageChops.difference(mask, plane)
+        elif subpath.operation == 1:
+            mask = ImageChops.lighter(mask, plane)
+        elif subpath.operation == 2:
+            if first:
+                mask = ImageChops.invert(mask)
+            mask = ImageChops.subtract(mask, plane)
+        elif subpath.operation == 3:
+            if first:
+                mask = ImageChops.invert(mask)
+            mask = ImageChops.darker(mask, plane)
+        first = False
     return mask.crop(layer.bbox)
 
 
-def draw_solid_color_fill(image, setting):
-    from PIL import ImageDraw
-    color = tuple(int(x) for x in setting.get(b'Clr ').values())
-    draw = ImageDraw.Draw(image)
-    draw.rectangle((0, 0, image.width, image.height), fill=color)
+def _draw_subpath(subpath, width, height):
+    from PIL import Image
+    import aggdraw
+    mask = Image.new('L', (width, height), 0)
+    path = ' '.join(map(str, _generate_symbol(subpath, width, height)))
+    draw = aggdraw.Draw(mask)
+    brush = aggdraw.Brush(255)
+    symbol = aggdraw.Symbol(path)
+    draw.symbol((0, 0), symbol, None, brush)
+    draw.flush()
     del draw
+    return mask
 
 
-def draw_pattern_fill(image, psd, setting):
+def _generate_symbol(path, width, height, command='C'):
+    """Sequence generator for SVG path."""
+    if len(path) == 0:
+        return
+
+    # Initial point.
+    yield 'M'
+    yield path[0].anchor[1] * width
+    yield path[0].anchor[0] * height
+    yield command
+
+    # Closed path or open path
+    points = (zip(path, path[1:] + path[0:1]) if path.is_closed()
+              else zip(path, path[1:]))
+
+    # Rest of the points.
+    for p1, p2 in points:
+        yield p1.leaving[1] * width
+        yield p1.leaving[0] * height
+        yield p2.preceding[1] * width
+        yield p2.preceding[0] * height
+        yield p2.anchor[1] * width
+        yield p2.anchor[0] * height
+
+    if path.is_closed():
+        yield 'Z'
+
+
+def draw_solid_color_fill(image, setting, blend=True):
+    from PIL import Image, ImageDraw, ImageChops
+    color = tuple(int(x) for x in setting.get(b'Clr ').values())
+    canvas = Image.new(image.mode, image.size)
+    draw = ImageDraw.Draw(canvas)
+    draw.rectangle((0, 0, canvas.width, canvas.height), fill=color)
+    del draw
+    if blend:
+        canvas.putalpha(image.getchannel('A'))
+        _blend(image, canvas, (0, 0))
+    else:
+        image.paste(canvas)
+
+
+def draw_pattern_fill(image, psd, setting, blend=True):
+    """
+    Draw pattern fill on the image.
+
+    :param image: Image to be filled.
+    :param psd: :py:class:`PSDImage`.
+    :param setting: Descriptor containing pattern fill.
+    :param blend: Blend the fill or ignore. Effects blend.
+    """
+    from PIL import Image
     pattern_id = setting[b'Ptrn'][b'Idnt'].value.rstrip('\x00')
     pattern = psd._get_pattern(pattern_id)
     if not pattern:
         logger.error('Pattern not found: %s' % (pattern_id))
         return None
     panel = convert_pattern_to_pil(pattern, psd._record.header.version)
-    for left in range(0, image.width, panel.width):
-        for top in range(0, image.height, panel.height):
-            image.paste(panel, (left, top))
+
+    scale = setting.get(b'Scl ', 100) / 100.
+    if scale != 1.:
+        panel = panel.resize((
+            int(panel.width * scale),
+            int(panel.height * scale)
+        ))
+
+    opacity = int(setting.get(b'Opct', 100) / 100. * 255)
+    if opacity != 255:
+        panel.putalpha(opacity)
+
+    pattern_image = Image.new(image.mode, image.size)
+    mask = image.getchannel('A') if blend else Image.new('L', image.size, 255)
+
+    for left in range(0, pattern_image.width, panel.width):
+        for top in range(0, pattern_image.height, panel.height):
+            panel_mask = mask.crop(
+                (left, top, left + panel.width, top + panel.height)
+            )
+            pattern_image.paste(panel, (left, top), panel_mask)
+
+    if blend:
+        image.paste(_blend(image, pattern_image, (0, 0)))
+    else:
+        image.paste(pattern_image)
 
 
 def draw_gradient_fill(image, setting, blend=True):
@@ -313,6 +408,7 @@ def draw_gradient_fill(image, setting, blend=True):
 
     gradient_image = _apply_color_map(image.mode, setting.get(b'Grad'), Z)
     if blend:
+        gradient_image.putalpha(image.getchannel('A'))
         _blend(image, gradient_image, offset=(0, 0))
     else:
         image.paste(gradient_image)
@@ -345,13 +441,17 @@ def _apply_color_map(mode, grad, Z):
     scalar = {
         'RGB': 1.0, 'L': 2.55, 'CMYK': 2.55,
     }.get(mode, 1.0)
+
+    X = [stop.get(b'Lctn').value / 4096. for stop in stops]
+    Y = [
+        tuple(int(scalar * x.value) for x in stop.get(b'Clr ').values())
+        for stop in stops
+    ]
+    if len(stops) == 1:
+        X = [0., 1.]
+        Y = [Y[0], Y[0]]
     G = interpolate.interp1d(
-        [stop.get(b'Lctn').value / 4096. for stop in stops],
-        [
-            tuple(int(scalar * x.value) for x in stop.get(b'Clr ').values())
-            for stop in stops
-        ],
-        axis=0, fill_value='extrapolate'
+        X, Y, axis=0, fill_value='extrapolate'
     )
     pixels = G(Z).astype(np.uint8)
     if pixels.shape[-1] == 1:
@@ -360,10 +460,13 @@ def _apply_color_map(mode, grad, Z):
     image = Image.fromarray(pixels, mode.rstrip('A'))
     if b'Trns' in grad and mode.endswith('A'):
         stops = grad.get(b'Trns')
+        X = [stop.get(b'Lctn').value / 4096 for stop in stops]
+        Y = [stop.get(b'Opct').value * 2.55 for stop in stops]
+        if len(stops) == 1:
+            X = [0., 1.]
+            Y = [Y[0], Y[0]]
         G_opacity = interpolate.interp1d(
-            [stop.get(b'Lctn').value / 4096 for stop in stops],
-            [stop.get(b'Opct').value * 2.55 for stop in stops],
-            axis=0, fill_value='extrapolate'
+            X, Y, axis=0, fill_value='extrapolate'
         )
         alpha = G_opacity(Z).astype(np.uint8)
         image.putalpha(Image.fromarray(alpha, 'L'))