+++ tests/psd_tools/api/test_composer.py
@@ -39,6 +39,23 @@ def test_compose_quality(filename):
     assert _calculate_hash_error(preview, rendered) <= 0.1
 
 
+@pytest.mark.parametrize(("filename",), [
+    ('path-operations/combine.psd',),
+    ('path-operations/exclude-first.psd',),
+    ('path-operations/exclude.psd',),
+    ('path-operations/intersect-all.psd',),
+    ('path-operations/intersect-first.psd',),
+    ('path-operations/subtract-all.psd',),
+    ('path-operations/subtract-first.psd',),
+    ('path-operations/subtract-second.psd',),
+])
+def test_compose_quality_rgb(filename):
+    psd = PSDImage.open(full_name(filename))
+    preview = psd.topil().convert('RGB')
+    rendered = psd.compose(force=True).convert('RGB')
+    assert _calculate_hash_error(preview, rendered) <= 0.1
+
+
 @pytest.mark.parametrize('filename', [
     'smartobject-layer.psd',
     'type-layer.psd',